v3.0
Complete and total revamp of the extension!

Notable changes:
- Single most exciting part, I've finally been able to resolve #4's issue with disconnected file inputs (and subsequently, the disconnected input in #13). Both of these inputs should work.
- Entirely revamped the extension, cutting ~250 lines of code out and using a slightly different method of injecting the script into webpages (notably for iframes and disconnected file inputs). The code is also much cleaner, and all unused code has been removed.
- Entirely redid how I identify file inputs. In theory, should be extremely reliable and less case/DOM-dependent.

This is what I would consider a beta version. It hasn't been extensively tested on a great number of websites (its only been tested on websites that have been previously known to not work, from various issues).

There is also a known regression (also from #13), in which the "quick-paste" feature no longer works. I couldn't quite find a way to work it into this release, but it'll be re-added in the next one.